requirements.txt
-e libs/core
-e libs/studio
fastapi==0.111.1
pydantic==2.8.2
pyinstaller==6.9.0
pystray==0.19.5
pyyaml==6.0.1
langchain==0.2.14
langchain-openai==0.1.22
langchain-groq==0.1.9
langchain-aws==0.1.17
pytest-asyncio==0.24.0
jsonschema==4.23